refactoed functions into module for specific migration areas
